fix: 修复 JSONObject.put() 不支持 Float 类型的错误
修复崩溃问题：
- JSONObject.put(String, float) 方法不存在
- 将 Float 转换为 Double：temperature.toDouble()

错误日志：
java.lang.NoSuchMethodError: No virtual method put(Ljava/lang/String;F)Lorg/json/JSONObject;

这是导致应用启动崩溃的根本原因

+++ android/app/src/main/java/com/chenyi/agent/Kernel.kt
@@ -319,7 +319,7 @@ data class KernelConfig(
         json.put("api_key", apiKey)
         json.put("model", modelName)
         json.put("max_tokens", maxTokens)
-        json.put("temperature", temperature)
+        json.put("temperature", temperature.toDouble())  // Float 转 Double
         return json.toString()
     }
 }
